// frontend/src/api/models/UsersRoleRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UsersRoleRequest
 */
export interface UsersRoleRequest {
    /**
     * 
     * @type {string}
     * @memberof UsersRoleRequest
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof UsersRoleRequest
     */
    roleId?: string;
}

/**
 * Check if a given object implements the UsersRoleRequest interface.
 */
export function instanceOfUsersRoleRequest(value: object): value is UsersRoleRequest {
    return true;
}

export function UsersRoleRequestFromJSON(json: any): UsersRoleRequest {
    return UsersRoleRequestFromJSONTyped(json, false);
}

export function UsersRoleRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsersRoleRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'userId': json['userId'] == null ? undefined : json['userId'],
        'roleId': json['roleId'] == null ? undefined : json['roleId'],
    };
}

export function UsersRoleRequestToJSON(value?: UsersRoleRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'userId': value['userId'],
        'roleId': value['roleId'],
    };
}