// frontend/src/api/models/UsersRoleResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserDTO } from './UserDTO';
import {
    UserDTOFromJSON,
    UserDTOFromJSONTyped,
    UserDTOToJSON,
} from './UserDTO';
import type { RoleResponse } from './RoleResponse';
import {
    RoleResponseFromJSON,
    RoleResponseFromJSONTyped,
    RoleResponseToJSON,
} from './RoleResponse';

/**
 * 
 * @export
 * @interface UsersRoleResponse
 */
export interface UsersRoleResponse {
    /**
     * 
     * @type {UserDTO}
     * @memberof UsersRoleResponse
     */
    user?: UserDTO;
    /**
     * 
     * @type {RoleResponse}
     * @memberof UsersRoleResponse
     */
    role?: RoleResponse;
}

/**
 * Check if a given object implements the UsersRoleResponse interface.
 */
export function instanceOfUsersRoleResponse(value: object): value is UsersRoleResponse {
    return true;
}

export function UsersRoleResponseFromJSON(json: any): UsersRoleResponse {
    return UsersRoleResponseFromJSONTyped(json, false);
}

export function UsersRoleResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsersRoleResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'user': json['user'] == null ? undefined : UserDTOFromJSON(json['user']),
        'role': json['role'] == null ? undefined : RoleResponseFromJSON(json['role']),
    };
}

export function UsersRoleResponseToJSON(value?: UsersRoleResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'user': UserDTOToJSON(value['user']),
        'role': RoleResponseToJSON(value['role']),
    };
}